Extract common code to a helper method
